Add basic readme and diagram of strava API interactions.
The main body of work in this commit is a link to a diagram which fleshes
out how the strava interactions are going to work. The Design.md document
also includes details about all the jobs that need to be implemented and
information about when they get queued and what happens if they succeed
or fail.