Add a job to fetch rides
We want to be able to fetch all the rides for a ride, no matter what
the current state of the rides associated with them looks like.

This means we need to support both fetching rides past the
most recent ride date we have and before the oldest ride date we have.

We can use the strava ruby client to iterate through ride pages and then
create the appropriate rides